requirements.txt
# Flask e extensões
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5

# Banco de dados
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Autenticação e segurança
Werkzeug==3.0.1
python-dotenv==1.0.0

# PDF
reportlab==4.0.7
weasyprint==60.1

# HTTP requests (Evolution API)
requests==2.31.0

# Utilitários
python-dateutil==2.8.2
